Finding missing DOIs
Missing DOIs are the result of changes of paper titles as well as minor text formatting issues. We can address this issue by relaxing the similarity function we use to find papers in the ACL anthology. I suggest that we use a boolean column to flag these papers (not yet implemented) and verify them manually at the end.